프론트엔드 개편 및 비로그인 유저 이용불가 추가

+++ src/main/java/com/springweek2/springweek2/service/PostService.java
@@ -1,13 +1,12 @@
 package com.springweek2.springweek2.service;
 
 import com.springweek2.springweek2.dto.PostRequestDto;
-import com.springweek2.springweek2.dto.SignupRequestDto;
 import com.springweek2.springweek2.model.Post;
 import com.springweek2.springweek2.repository.PostRepository;
 import com.springweek2.springweek2.repository.UserRepository;
+import com.springweek2.springweek2.security.UserDetailsImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import javax.transaction.Transactional;
 import java.util.List;
@@ -39,20 +38,26 @@ public class PostService {
     }
 
     @Transactional
-    public Long updatePost(Long id, PostRequestDto requestDto) {
-        Post post = postRepository.findById(id).orElseThrow(
-                () -> new NullPointerException("선택한 게시글이 존재하지 않습니다.")
-        );
-        String contents = requestDto.getContents();
-        String title = requestDto.getTitle();
-        post.setContents(contents);
-        post.setTitle(title);
-        postRepository.save(post);
-        return id;
+    public Long updatePost(Long id, PostRequestDto requestDto, UserDetailsImpl userDetails) {
+        if (userDetails.getUser().getUsername().equals(postRepository.getById(id).getUsername())) {
+            Post post = postRepository.findById(id).orElseThrow(
+                    () -> new NullPointerException("선택한 게시글이 존재하지 않습니다.")
+            );
+            String contents = requestDto.getContents();
+            String title = requestDto.getTitle();
+            post.setContents(contents);
+            post.setTitle(title);
+            postRepository.save(post);
+            return id;
+        }
+        return -1L;
     }
 
-    public Long deletePost(Long id) {
-        postRepository.deleteById(id);
-        return id;
+    public Long deletePost(Long id, UserDetailsImpl userDetails) {
+        if (userDetails.getUser().getUsername().equals(postRepository.getById(id).getUsername())) {
+            postRepository.deleteById(id);
+            return id;
+        }
+        return -1L;
     }
 }
